Fixed comments for part of frontend and split up components and layouts in correct folders
