+++ frontend/src/App.tsx
@@ -3,10 +3,10 @@ import { Routes, Route, Navigate } from "react-router-dom";
 import Login from "./pages/Login";
 import Home from "./pages/Home";
 import Verify from "./pages/Verify";
-import LearningLayout from './layouts/LearningLayout';
-import PageContent from './components/PageContent';
-import FlashcardLayout from './layouts/FlashcardLayout'; // Import new layout
-import CalendarView from './components/CalendarView'
+import LearningLayout from './layouts/LearningEnvironment/LearningLayout';
+import PageContent from './components/LearningEnvironment/PageContent';
+import FlashcardLayout from './layouts/Flashcards/FlashcardLayout';
+import CalendarView from './components/Calendar/CalendarView'
 import LoginSuccessPage from './pages/LoginSuccessPage'; 
 
 function App() {
@@ -30,11 +30,9 @@ function App() {
 
       <Route
           path="/learning"
-          // Apply the SAME protection logic as /home if needed
-          // If learning is public, just use: element={<LearningLayout />}
+          // Apply the SAME protection logic as /home
           element={token ? <LearningLayout /> : <Navigate to="/login" />}
         >
-          {/* Nested routes render inside LearningLayout's <Outlet /> */}
           <Route index element={<p>Velkommen til læringsområdet!</p>} />
           <Route path=":pageId" element={<PageContent />} />
       </Route>
