+++ frontend/src/components/Flashcards/FlashcardSideNav.tsx
@@ -1,8 +1,8 @@
 // src/components/FlashcardSideNav.tsx
 import { useState, useEffect } from 'react';
 import { NavLink } from 'react-router-dom'; // useLocation can be useful
-import { fetchFlashcardCollections } from '../services/ApiService'; // Adjust path if needed
-import type { FlashcardCollectionSummaryDto } from '../types/flashcardTypes'; // Adjust path if needed
+import { fetchFlashcardCollections } from '../../services/ApiService'; // Adjust path if needed
+import type { FlashcardCollectionSummaryDto } from '../../types/flashcardTypes'; // Adjust path if needed
 import './FlashcardSideNav.css'; // Create this CSS file
 
 function FlashcardSideNav() {
