+++ frontend/src/components/LearningEnvironment/PageContent.css
